requirements.txt
aiohttp==3.8.4; python_version >= "3.7"
aiosignal==1.3.1; python_version >= "3.7"
altair==5.0.1; python_version >= "3.7"
anyio==3.6.2; python_full_version >= "3.6.2" and python_version >= "3.7"
async-timeout==4.0.2; python_version >= "3.7"
attrs==23.1.0; python_version >= "3.7"
cachetools==5.3.0; python_version >= "3.7" and python_version < "4.0"
certifi==2023.5.7; python_version >= "3.7"
charset-normalizer==3.1.0; python_full_version >= "3.7.0" and python_version >= "3.7"
click==8.1.3; python_version >= "3.7"
colorama==0.4.6; python_version >= "3.7" and python_full_version < "3.0.0" and platform_system == "Windows" or platform_system == "Windows" and python_version >= "3.7" and python_full_version >= "3.7.0"
contourpy==1.0.7; python_version >= "3.8"
cycler==0.11.0; python_version >= "3.8"
fastapi==0.95.2; python_version >= "3.7"
ffmpy==0.3.0; python_version >= "3.7"
fonttools==4.39.4; python_version >= "3.8"
frozenlist==1.3.3; python_version >= "3.7"
fsspec==2023.5.0; python_version >= "3.8"
google-api-core==2.11.0; python_version >= "3.7"
google-auth==2.19.0; python_version >= "3.7"
google-cloud-vision==3.4.1; python_version >= "3.7"
googleapis-common-protos==1.59.0; python_version >= "3.7" and (python_version >= "3.11" or python_version >= "3.11")
gradio==3.14.0; python_version >= "3.7"
grpcio-status==1.55.0; python_version >= "3.7" and (python_version >= "3.11" or python_version >= "3.11")
grpcio==1.55.0; python_version >= "3.7" and (python_version >= "3.11" or python_version >= "3.11")
h11==0.14.0; python_version >= "3.7"
httpcore==0.17.2; python_version >= "3.7"
httpx==0.24.1; python_version >= "3.7"
idna==3.4; python_full_version >= "3.6.2" and python_version >= "3.7"
importlib-resources==5.12.0; python_version < "3.10" and python_version >= "3.8"
jinja2==3.1.2; python_version >= "3.7"
jsonschema==4.17.3; python_version >= "3.7"
kiwisolver==1.4.4; python_version >= "3.8"
linkify-it-py==2.0.2; python_version >= "3.7"
markdown-it-py==2.2.0; python_version >= "3.7"
markupsafe==2.1.2; python_version >= "3.7"
matplotlib==3.7.1; python_version >= "3.8"
mdit-py-plugins==0.3.5; python_version >= "3.7"
mdurl==0.1.2; python_version >= "3.7"
multidict==6.0.4; python_version >= "3.7"
numpy==1.24.3
orjson==3.8.14; python_version >= "3.7"
packaging==23.1; python_version >= "3.8"
pandas==2.0.1; python_version >= "3.8"
pillow==9.5.0; python_version >= "3.8"
proto-plus==1.22.2
protobuf==4.23.2
pyasn1-modules==0.3.0; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.7"
pyasn1==0.5.0; python_version >= "3.7" and python_full_version < "3.0.0" and python_version < "4" or python_full_version >= "3.6.0" and python_version >= "3.7" and python_version < "4"
pycryptodome==3.18.0; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.7"
pydantic==1.10.8; python_version >= "3.7"
pydub==0.25.1; python_version >= "3.7"
pyparsing==3.0.9; python_full_version >= "3.6.8" and python_version >= "3.8"
pyrsistent==0.19.3; python_version >= "3.7"
python-dateutil==2.8.2; python_version >= "3.8" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.8" \
    --hash=sha256:0123cacc1627ae19ddf3c27a5de5bd67ee4586fbdd6440d9748f8abb483d3e86 \
    --hash=sha256:961d03dc3453ebbc59dbdea9e4e11c5651520a876d0f4db161e8674aae935da9
python-dotenv==1.0.0; python_version >= "3.8"
python-multipart==0.0.6; python_version >= "3.7"
pytz==2023.3; python_version >= "3.8"
pyyaml==6.0; python_version >= "3.7"
requests==2.31.0; python_version >= "3.7"
rsa==4.9; python_version >= "3.7" and python_version < "4"
setuptools-scm==7.1.0; python_version >= "3.8"
six==1.16.0; python_version >= "3.8" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.8" \
    --hash=sha256:8abb2f1d86890a2dfb989f9a77cfcfd3e47c2a354b01111771326f8aa26e0254 \
    --hash=sha256:1e61c37477a1626458e36f7b1d82aa5c9b094fa4802892072e49de9c60c4c926
sniffio==1.3.0; python_full_version >= "3.6.2" and python_version >= "3.7"
starlette==0.27.0; python_version >= "3.7"
tomli==2.0.1; python_version < "3.11" and python_version >= "3.8" \
    --hash=sha256:939de3e7a6161af0c887ef91b7d41a53e7c5a1ca976325f429cb46ea9bc30ecc \
    --hash=sha256:de526c12914f0c550d15924c62d72abc48d6fe7364aa87328337a31007fe8a4f
toolz==0.12.0; python_version >= "3.7"
typing-extensions==4.6.2; python_version < "3.10" and python_version >= "3.8"
tzdata==2023.3; python_version >= "3.8"
uc-micro-py==1.0.2; python_version >= "3.7"
urllib3==1.26.16; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.7"
uvicorn==0.22.0; python_version >= "3.7"
websockets==11.0.3; python_version >= "3.7"
yarl==1.9.2; python_version >= "3.7"
zipp==3.15.0; python_version < "3.10" and python_version >= "3.8"
